Move delta token example into original delta examples
